% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moirasays.R
\docType{data}
\name{moira}
\alias{moira}
\title{Select Random Text}
\format{
An object of class \code{character} of length 101.
}
\usage{
moira
}
\arguments{
\item{text_list}{A character vector of text items.}

\item{n}{Number of items to select.}

\item{replace}{Sample with replacement? (Default: FALSE)}
}
\value{
A randomly selected character vector.
}
\description{
Select Random Text
}
\examples{
my_texts <- c("Hello", "World", "R", "Package")
select_random(my_texts, n = 2)

}
\keyword{datasets}
