% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moirasays.R
\docType{data}
\name{moira_text}
\alias{moira_text}
\title{Select Random Text}
\format{
An object of class \code{character} of length 101.
}
\usage{
moira_text
}
\arguments{
\item{text_list}{A character vector of text items.}

\item{n}{Number of items to select.}

\item{replace}{Sample with replacement? (Default: FALSE)}
}
\value{
A randomly selected character vector.
}
\description{
Select Random Text
}
\keyword{datasets}
