% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moirasays.R
\name{moirasays}
\alias{moirasays}
\title{Select Moira}
\usage{
moirasays(text_list = moira, n = 1, replace = FALSE)
}
\arguments{
\item{text_list}{A character vector of text items.}

\item{n}{Number of items to select.}

\item{replace}{Sample with replacement? (Default: FALSE)}
}
\value{
A randomly selected character vector.
}
\description{
Select Moira
}
\examples{
my_texts <- c("Hello", "World", "R", "Package")
select_random(my_texts, n = 2)

}
