% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moirasays.R
\name{moirasays}
\alias{moirasays}
\title{Select Moira Phrases}
\usage{
moirasays(text_list = moira_list, n = 1, replace = FALSE)
}
\arguments{
\item{text_list}{A character vector of text items, included with this package}

\item{n}{Number of items to select.}

\item{replace}{Sample with replacement? (Default: FALSE)}
}
\value{
A randomly selected character vector.
}
\description{
Select Moira Phrases
}
